Add meson build system
[mesonbuild] is a new (not so new already) build system that is
extensively used on the linux desktop.

It would be fantastic to have support for building fdkaac with Meson
in upstream fdkaac, as it would facilitate builds of GStreamer via
Meson's subproject support.

Cross-compilation is of course also supported, but I have not tested
that yet. Meson provides out-of-the-box support for building native
binaries and cross binaries in the same build, which seems like
something that could be useful for fdkaac.

This should be on par with the CMake build definition

[mesonbuild]:https://mesonbuild.com/