github: Build the whole multiplatform docker image with Dockerfile.toolchain
This reverts commit 8f421127194ea0ca2325bbe6648a783612718619,
and fixes building the multiplatform image in a simpler way.

Instead of doing two separate builds, with separate Dockerfiles,
just do one multiplatform build of one Dockerfile, which
docker runs serially, for each one of the included architectures,
just packaging the prebuilt toolchains in the docker image.

This makes it more feasible to get a PGO optimized compiler in
the docker images, without needing to redo all the PGO build
stages again here.