Improve "spatial neighbour" context derivation.
On overhangs (bottom, right), the edges might not be available, and
therefore we can't assume there's always an even number of edges
available. For example, we might be both at a superblock boundary
(above not available) as well as have overhanging edges on the left
(bottom-left not available), in which case we have only one context
(left).