msac: fill dif with 1s instead of 0s.
This is 1-2 instructions more expensive (which is why it was the
way it was to begin with), but I don't get the same entropy decoding
behaviour in the bypass coding without this. My hunch is that long
bypass symbols (unary/bools can read up to 32 bits) can compare
against unfilled bits, but I haven't verified that.