ci: Add a Github Actions CI job for testing building with MSVC
Test building for x86 and x64 with MSVC on x86_64 Windows,
and for arm64 on arm64 Windows. Unfortunately, the current versions
of Windows no longer support running 32 bit ARM binaries, and
the latest versions of WinSDK no longer contain support for building
for it.

To cover this gap, also test building on Linux, with msvc-wine,
where we can request installing a specific older version of the
WinSDK and MSVC.

Also test building with clang-cl, for one architecture.

Also test cross compiling with MSVC for arm64, on x86_64. This
is special because in this setup, there's no working host
compiler (that can build executables to run during the build),
only a cross compiler.