checkasm/test: fall back to naive loop for low iteration counts
CHECKASM_PERF_BENCH_ASM always unrolls 32 loop iterations; this is bad when
measuring very slow functions because it means we only collect data about a
single iteration count (effectively 32). It also leads to vastly overshooting
the time budget.

It's much better for both the linear regression and the overall runtime to
collect data points as evenly as possible, so fall back to the naive loop when
the number of iterations is so low.