Improve/simplify handling of intra blocks in get_comp_ctx().
The previous implementation didn't handle intra blocks correctly in
case of both refs being non-compound (refdir[refa1] ^ refdir[refb1])
or in case of intrabc-blocks if the other ref is missing or compound.

Rather than complicate the function further, prepend intra as a new
entry at the start of refdir[], and add a special case for intrabc
where needed (if the other ref is compound or missing).