Handle cases where warp-contexting goes outside a block's bounds.
The relevant AVM functions for this are av1_is_warp_causal_allowed()
and inter_warpmv_mode_ctx(). At SB boundaries, these functions limit
resolution of the "top" context to 8x8, but because of how that's
implemented, this can make the context be outside the current block's
positions. For example, for a block at y=128,x=225,w=2, the "top" may
be at y=128-1,x=224 (instead of x=225 or x=226). The practical problem
we face here is that the "top" x=224 value would have been overwritten
by our left-neighbour already, so it does not exist anymore.

Since this only affects 3 values (ref[0], ref[1] and motion_mode) and
only at SB boundaries, we can add a "backup" context at SB boundaries
and use that in these situations. I've filed a bug upstream to ask if
this behaviour is intentional (#1091).