Add tip frames.
They are currently not yet reconstructed (that will have to be hooked up
separately), but they are added to the reference list and the CDFs are
processed as needed also. There is no bitstream data associated with these
frames, so we only need reconstruction (what the threaded code calls pass=2).