Increase size of levels[] by 1 byte to prevent overwrite into txtp_map[].
When decoding IDTX, we pad width *and* height by 1 and memset(0)
the whole area (33x33), as opposed to 1D transforms, where we have
2 bytes padding in 1 dimension (to a max of 32x34, i.e. 1 byte less).
But when the new forward IDTX coding was added, the levels[] array
was not increased. This may result in a 1-byte overwrite into
txtp_map[], which would break chroma/inter txtp inferral. Increasing
levels[] by 1 byte fixes that.