Skip failing selftests if we don't have a working signal handler
In particular, the Windows UWP configuration doesn't have a working
signal handler. For that case, make checkasm_should_fail return a
value to indicate whether we should proceed with the expectedly
failing selftests or not.

This can also be used for the perf init code for asm timers, which
may rely on signal handlers for fallbacks. If we don't have a working
signal handler, skip trying the asm timers.