msac: Improve refill logic
Perform an early refill if necessary in decode_bools_bypass() the
same way it's already done in decode_unary_bypass(). This allows
us to reduce the minimum number of bits kept in the ec window
to match AV1 which reduces the total number of refills required.

Also prefer using unsigned compares for the refill condition
to avoid redundant refill attempts when reaching EOB.